Simplify AsyncIO part 1: Remove ThreadEventQueue.h
